Only download the simulated eye tracking model if we want to use it

+++ docker/image_processing/entrypoint.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 
-if [ ! -f "/var/source/saved_model.pb" ]
+if echo "$@" | grep -q "simulated_eye_tracking"
   then
-    wget -O /var/source/saved_model.pb 'https://huggingface.co/JackVines/ds_saliency_inference/resolve/main/app/saved_model.pb'
+    if [ ! -f "/var/source/saved_model.pb" ]
+      then
+        wget -O /var/source/saved_model.pb 'https://huggingface.co/JackVines/ds_saliency_inference/resolve/main/app/saved_model.pb'
+    fi
 fi
-
 pip install --root-user-action=ignore --disable-pip-version-check tensorflow opencv-python-headless | grep -v "^Requirement already satisfied"
 
 # Run the given command
